Move all classes to their own file
