+++ main/src/addins/MonoDevelop.UnitTesting/EventArgs/TestSessionEventArgs.cs
@@ -1,5 +1,5 @@
 //
-// IResultsStoreSerializer.cs
+// TestSessionEventArgs.cs
 //
 // Author:
 //   Lluis Sanchez Gual
@@ -29,28 +29,27 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.Xml.Serialization;
+using System.Threading;
+
 using MonoDevelop.Core;
+using MonoDevelop.Core.Execution;
+using Mono.Addins;
+using MonoDevelop.Projects;
+using MonoDevelop.Ide.Gui;
+using MonoDevelop.Ide;
+using System.Threading.Tasks;
+using System.Linq;
+using MonoDevelop.Ide.TypeSystem;
+using System.IO;
+using MonoDevelop.Ide.Gui.Components;
 
 namespace MonoDevelop.UnitTesting
 {
-	
-	/// <summary>
-	/// Encapsulates serialization/deserialization logic
-	/// </summary>
-	public interface IResultsStoreSerializer
+
+	public class TestSessionEventArgs: EventArgs
 	{
-		/// <summary>
-		/// Serialize the record into the specified path.
-		/// </summary>
-		void Serialize(string filePath, TestRecord testRecord);
-		
-		/// <summary>
-		/// Deserialize the TestRecord from the sepcified path if possible.
-		/// Return null if deserialization is impossible.
-		/// </summary>
-		TestRecord Deserialize(string filePath);
+		public AsyncOperation Session { get; set; }
+		public UnitTest Test { get; set; }
 	}
+	
 }
