+++ main/src/addins/MonoDevelop.UnitTesting/Services/IUnitTestMarkers.cs
@@ -2,9 +2,9 @@
 // IUnitTestMarkers.cs
 //
 // Author:
-//       Anthony Stivers <tony.stivers@gmail.com>
+//       Mike Krüger <mkrueger@xamarin.com>
 //
-// Copyright (c) 2016 Anthony Stivers
+// Copyright (c) 2013 Xamarin Inc. (http://xamarin.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
