+++ main/src/addins/MonoDevelop.UnitTesting/Services/UnitTestService.cs
@@ -389,72 +389,5 @@ namespace MonoDevelop.UnitTesting
 		/// </summary>
 		public static event EventHandler<TestSessionEventArgs> TestSessionStarting;
 	}
-	
-
-
-	class TestSession: AsyncOperation
-	{
-		UnitTest test;
-		TestMonitor monitor;
-		MonoDevelop.Projects.ExecutionContext context;
-		TestResultsPad resultsPad;
-
-		public TestSession (UnitTest test, MonoDevelop.Projects.ExecutionContext context, TestResultsPad resultsPad, CancellationTokenSource cs)
-		{
-			this.test = test;
-			if (context != null)
-				this.context = new Projects.ExecutionContext (context.ExecutionHandler, new CustomConsoleFactory (context.ConsoleFactory, cs), context.ExecutionTarget);
-			CancellationTokenSource = cs;
-			this.monitor = new TestMonitor (resultsPad, CancellationTokenSource);
-			this.resultsPad = resultsPad;
-			resultsPad.InitializeTestRun (test, cs);
-			Task = new Task ((Action)RunTests);
-		}
-		
-		public Task Start ()
-		{
-			Task.Start ();
-			return Task;
-		}
-
-		void RunTests ()
-		{
-			try {
-				UnitTestService.ResetResult (test);
-
-				TestContext ctx = new TestContext (monitor, resultsPad, context, DateTime.Now);
-				test.Run (ctx);
-				test.SaveResults ();
-			} catch (Exception ex) {
-				LoggingService.LogError (ex.ToString ());
-				monitor.ReportRuntimeError (null, ex);
-			} finally {
-				monitor.FinishTestRun ();
-			}
-		}
-	}
-
-	public class TestSessionEventArgs: EventArgs
-	{
-		public AsyncOperation Session { get; set; }
-		public UnitTest Test { get; set; }
-	}
-
-	class CustomConsoleFactory : OperationConsoleFactory
-	{
-		OperationConsoleFactory factory;
-		CancellationTokenSource cancelSource;
-
-		public CustomConsoleFactory (OperationConsoleFactory factory, CancellationTokenSource cs)
-		{
-			this.factory = factory;
-			cancelSource = cs;
-		}
-
-		protected override OperationConsole OnCreateConsole (CreateConsoleOptions options)
-		{
-			return factory.CreateConsole (options.WithBringToFront (false)).WithCancelCallback (cancelSource.Cancel);
-		}
-	}
 }
